BE/pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiomysql>=0.2.0",
    "fastapi>=0.115.12",
    "google-genai>=1.14.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.5.0",
    "langsmith>=0.4.4",
    "mcp[cli]>=1.10.1",
    "notion-client>=2.4.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "qdrant-client>=1.14.2",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pre-commit>=4.2.0",
]
